Changed Primary Camera and Added Camera Pan
Added a new cinematic camera to the scene. The level blueprint sets the new main view target with the new camera and also adds a constant transform to the camera's Y axis using a "camera speed" float variable. This transform is applied to the camera using the tick event on every frame