Incomplete Shooting Logic
Adding section of nodes to the third person character blueprint for spawning a projectile based on the cursor position. The projectile is also a basic blueprint with some simple move logic. Spawning still needs to be changed to ignore depth (projectiles will fly into foreground/background based on cursor position) and projectile mesh needs to be swapped.